Refactored to make each about field importable
