+++ plugins/catalog-react/src/utils/getEntityRelations.ts
@@ -16,8 +16,8 @@
 
 import {
   Entity,
-  CompoundEntityRef,
   parseEntityRef,
+  stringifyEntityRef,
 } from '@backstage/catalog-model';
 
 // TODO(freben): This should be returning entity refs instead
@@ -30,7 +30,7 @@ export function getEntityRelations(
   entity: Entity | undefined,
   relationType: string,
   filter?: { kind: string },
-): CompoundEntityRef[] {
+): string[] {
   let entityNames =
     entity?.relations
       ?.filter(r => r.type === relationType)
@@ -44,5 +44,5 @@ export function getEntityRelations(
     );
   }
 
-  return entityNames;
+  return entityNames.map(stringifyEntityRef);
 }
