--- plugins/catalog/src/components/AboutCard/EditMetadataButton.tsx
@@ -1,40 +1,0 @@
-/*
- * Copyright 2022 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { ANNOTATION_EDIT_URL } from '@backstage/catalog-model';
-import { Link } from '@backstage/core-components';
-import { useEntity } from '@backstage/plugin-catalog-react';
-import { IconButton } from '@material-ui/core';
-import EditIcon from '@material-ui/icons/Edit';
-import React from 'react';
-
-export function EditMetadataButton() {
-  const { entity } = useEntity();
-
-  const entityMetadataEditUrl =
-    entity.metadata.annotations?.[ANNOTATION_EDIT_URL];
-
-  return (
-    <IconButton
-      component={Link}
-      aria-label="Edit"
-      disabled={!entityMetadataEditUrl}
-      title="Edit Metadata"
-      to={entityMetadataEditUrl ?? '#'}
-    >
-      <EditIcon />
-    </IconButton>
-  );
-}
